Skip .git subdirs.  Don't print defaulted paths
If a .git subdir is encountered during tree scan it is skipped.
A .git directory can still be scanned by specifying it as the
path to scan, but this is would simply verify that .git text
objects use LF endings.

If no path is specified on the command line the path to scan
defaults to './'  In this case the leading './' is not printed
when endings are reported.